backend/pyproject.toml
[project]
authors = [
  {name = "MakerBi", email = "biying@vectorvein.com"},
]
dependencies = [
    "pywebview>=4.1",
    "httpx>=0.22.0",
    "peewee>=3.16.2",
    "pypdf>=3.8.1",
    "mammoth>=1.5.1",
    "openpyxl>=3.1.2",
    "python-pptx>=0.6.21",
    "openai>=1.0.0",
    "python-docx>=0.8.11",
    "pandas>=2.0.1",
    "yagmail>=0.15.293",
    "qdrant-client>=1.1.3",
    "markdown2>=2.4.8",
    "tiktoken>=0.4.0",
    "pyinstaller>=5.11.0",
    "readability-lxml>=0.8.1",
    "markdownify>=0.11.6",
    "yt-dlp>=2023.3.4",
    "pycryptodome>=3.18.0",
    "clr-loader>=0.2.6",
    "pyecharts>=2.0.3",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
]
description = ""
license = {text = "CC BY-NC-SA"}
name = ""
requires-python = ">=3.9,<3.13"
version = "0.0.31"

[project.optional-dependencies]
mac = ["pyobjc"]

[tool.pdm.scripts]
build = "python build.py -v 0.0.31 -t p"
build-debug = "pyinstaller debug.spec --noconfirm"
build-front = "python build.py -t f"
dev.cmd = "python main.py"
dev.env = {VECTORVEIN_DEBUG = "1", VECTORVEIN_VERSION = "0.0.31"}
fullstack-dev.cmd = "python run_fullstack_dev.py"
fullstack-dev.env = {VECTORVEIN_DEBUG = "1", VECTORVEIN_VERSION = "0.0.31", "VITE_LOCAL" = "http://localhost:5173/"}
